Make commit.py more pythonic.
First, if you're only using a compiled regular expression once, at least use
it completely. Every compiled regular expression in python allows you to call
the re functions on it passing one less argument.

Second, if you find yourself constantly checking to make sure the length of an
object is greater than 0, you're doing it wrong. In python a non-empty object (a
list, dictionary, tuple, or string that is not [], {}, () or '' respectively) is
'True' or > 0 and an empty object is 0 or 'False'. This simplifies the code
greatly.

In reality, lines 71-80 could probably be simplified, but it isn't exactly
important.