A whole bunch of simple changes.
type(var) is type is deprecated in newer versions of Python if I remember
correctly. Forward compatibility suggests isinstance(var, (type1, type2,
...)) be used in it's place where (type1, type2, ...) is a tuple of options
for the object type var can be.

len(list/tuple/string) - 1 used as an index can be simplified to var[-1]

string's have a built-in method called join which takes a list. Any string
will work and it will only place the string in-between the elements of the
list so one doesn't have to concern themselves with using a for-loop to
concatenate items as a string and then remove the last iteration's joining
string. This just required a quick list comprehension which python eats for
breakfast.

Also one doesn't have to call the object's representation function using
var.__repr__(), repr(var) calls that for you and is easier to type and read.

A few cosmetic changes included as well, otherwise I couldn't tell you if
the tests failed or not. I don't know what they're supposed to look off the
top of my head. They do look like they work though.