major changes to the keyword tagging to support mappig and apicall tags, as well as a proper graph for tags (no visualization quite yet).
Still, this whole thing is crap.  I'm applying a 'keywordtype' to an edge in the database, but applying it to a tag in the code.  There's comments sprinkled around about how this has several corner cases where it just flat out doesn't work.  Also the cycle detection is hackish.

Evenually, this will have to be rewritten again, to handle the weird cases I'm ignoring now by just not entering them in the database.  But for now, it should be robust enough to support a beta.