This is a particurally giant commit.
rss.py & browsedb.py
    chages things around to do a fulltext search by default
commit.py
    adds a function to test a string to see if it's present in any of the diffs of the commit
databaseQueries.py
    function renames, but also changes findByKeywordsAndFulltext to do fulltext searching, using eval()! Scary!
keywordsfilter.py
    Changes the tree around so it can be used to create a SQL where clause or a python string that can be eval()-ed
    Tree.__init__ determines if a word is a tag or a fulltext search item and tags it accordingly
    Test cases had to be updated acordingly
    Strips illegal characters in _stripIllegalCharacters as a second check (the first is the lexer) against a person being able to inject single or double quotes and gain code exec

We start with a string like 'sha256 or (talon and doxygen)'.  sha256 and doxygen are tags - we know ahead of time whether or not they're matched in the diff.  Talon is not a keyword, we have to search for it fulltext.

We create a sql query that has a WHERE clause that looks like this:

    'slackware' IN (SELECT ck.keyword FROM commitkeyword_tbl as ck WHERE ck.commitid = c.id)
    or  (
        1 = 1
        and
        'doxygen' IN (SELECT ck.keyword FROM commitkeyword_tbl as ck WHERE ck.commitid = c.id)
        )

This preserves the user's orders, but any 'unknown' quantity - any fulltext search - is replaced with a 1=1 to assume all records match.

We get all these commits back from the database, convert them to proper objects, and store them in an array.  Then we create a python test that looks like this:

    'slackware' in c.dbkeywords or (c.testFulltext('tag1') and 'doxygen' in c.dbkeywords)

This is the same query the user made, but now we can test the unknown quantities AND still preserve the same complicated heirarchical and/or logic dictated by the user.  This test is eval()-ed to determine whether or not the commit is interesting.